+++ src/services/reducers/ingredients.jsx
@@ -1,0 +1,64 @@
+import {
+  GET_INGREDIENTS_REQUEST,
+  GET_INGREDIENTS_SUCCESS,
+  GET_INGREDIENTS_FAILED,
+  SELECT_INGREDIENT,
+  UNSELECT_INGREDIENT,
+  SELECT_MEAL
+} from '../actions/ingredients';
+//--------------------------------------------------------------------------------
+
+const ingredientsInitialState = {
+  ingredients: [],
+  ingredientsRequest: false,
+  ingredientsFailed: false,
+  selectedIngredient: null,
+  selectedMeal: 'bun'
+};
+
+export const ingredientsReducer = (state = ingredientsInitialState, action) => {
+  switch (action.type) {
+    case GET_INGREDIENTS_REQUEST: {
+      return {
+        ...state,
+        ingredientsFailed: false,
+        ingredientsRequest: true
+      };
+    }
+    case GET_INGREDIENTS_SUCCESS: {
+      return {
+        ...state,
+        ingredientsFailed: false,
+        ingredientsRequest: false,
+        ingredients: action.payload.ingredients
+      };
+    }
+    case GET_INGREDIENTS_FAILED: {
+      return {
+        ...ingredientsInitialState,
+        ingredientsFailed: true
+      };
+    }
+    case SELECT_INGREDIENT: {
+      return {
+        ...state,
+        selectedIngredient: action.payload.ingredient
+      };
+    }
+    case UNSELECT_INGREDIENT: {
+      return {
+        ...state,
+        selectedIngredient: null
+      };
+    }
+    case SELECT_MEAL: {
+      return {
+        ...state,
+        selectedMeal: action.payload.mealName
+      };
+    }
+    default: {
+      return state;
+    }
+  }
+};
