+++ src/services/reducers/order.jsx
@@ -1,0 +1,46 @@
+import {
+  GET_ORDER_NUMBER_REQUEST,
+  GET_ORDER_NUMBER_SUCCESS,
+  GET_ORDER_NUMBER_FAILED,
+  DELETE_ORDER
+} from '../actions/order';
+//--------------------------------------------------------------------------------
+
+const orderInitialState = {
+  order: null,
+  orderNumberRequest: false,
+  orderNumberFailed: false
+};
+
+export const orderReducer = (state = orderInitialState, action) => {
+  switch (action.type) {
+    case GET_ORDER_NUMBER_REQUEST: {
+      return {
+        ...state,
+        orderNumberRequest: true
+      };
+    }
+    case GET_ORDER_NUMBER_SUCCESS: {
+      return {
+        ...state,
+        orderNumberRequest: false,
+        orderNumberFailed: false,
+        order: action.payload.order
+      };
+    }
+    case GET_ORDER_NUMBER_FAILED: {
+      return {
+        ...orderInitialState,
+        orderNumberFailed: true
+      };
+    }
+    case DELETE_ORDER: {
+      return {
+        ...state,
+        order: null
+      };
+    }
+    default:
+      return state;
+  }
+};
