+++ src/services/store.jsx
@@ -1,7 +1,6 @@
 import thunk from 'redux-thunk';
 import { applyMiddleware, compose, createStore } from 'redux';
-import { rootReducer } from './reducers/root-reducer';
-
+import { rootReducer } from './reducers';
 const composeEnhancers =
   typeof window === 'object' && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
     ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({})
